+++ src/memio.c
@@ -64,7 +64,7 @@ size_t mem_fread(void *buf, size_t size, size_t nmemb, MEMFILE *stream)
 	if (stream->mode != MODE_READ)
 	{
 		printf("not a read stream\n");
-		return -1;
+		return (size_t)-1;
 	}
 
 	// Trying to read more bytes than we have left?
@@ -112,7 +112,7 @@ size_t mem_fwrite(const void *ptr, size_t size, size_t nmemb, MEMFILE *stream)
 
 	if (stream->mode != MODE_WRITE)
 	{
-		return -1;
+		return (size_t)-1;
 	}
 	
 	// More bytes than can fit in the buffer?
